Renamed phazor_web/generated to components
