+++ phazor_web/src/components/info.rs
